FT_STRRCHR Modifications, not tested

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: sinkosi <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/24 11:42:49 by sinkosi           #+#    #+#             */
-/*   Updated: 2019/05/27 14:04:16 by sinkosi          ###   ########.fr       */
+/*   Updated: 2019/05/27 14:52:50 by sinkosi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchar(const char *s, int c)
 {
-	int i;
+	size_t i;
 
 	i = (ft_strlen(s) + 1);
 	while (i >= 0)
